Internal rework of RestRequestUtil.
Twitch made it so endpoint structures are no longer required to adhere
to the usual response structure, i.e they introduced a new format in the
follows data structure. The interal pipeline for making restrequests was
changed to accomodate any structure.

Re-tresting of all endpoints for strcuture changes is currently
happening.