Fixed a RateLimit bug. Updated /users/follows models. More testing.
There was a bug when converting the 'reset' header to a DateTime.
Instead of converting the 'Ratelimit-Reset' header, it was converting
the 'Ratelimit-Remaining' header. This is now fixed.

The Data and DataPage have been updated to now handle the top level
field 'total'.

Testing of the methods for /user/follows endpoint were testing and
perfomring as intended.