IRC Tags
The IRC tags dictionary has been replaces with a custom struct that behaves the same as a read-only dictionary. This should make it so that IRC messages are now laid out in continuous memory.

Fixed a couple bugs with converting tags to arrays that would return an array with a single default element instead of an empty array.