PubSub
Finished adding async connection method. Added initial support for events and handlers for complete messages. Added async send wrappers. Close frames are now handled properly, and the close status codes and reasons are now decoded and passed to the user (if applicable). Some refactoring was done to the payload data and frames to make more sense of how the data is handled.